refactor spaghetti MeshPath code into spaghetti.js instead of overloading humanPose/index.js
